Move run_all_trades to core and improve tests
