+++ tests/test_clean_import.py
@@ -12,6 +12,6 @@ def test_package_import_has_no_cli_side_effects(monkeypatch):
         "ai_trading.app",
         "ai_trading.main",
         "ai_trading.production_system",
-        "ai_trading.run_all_trades",
+        "ai_trading.core.run_all_trades",
     }
     assert heavy_modules.isdisjoint(sys.modules)
